# Added index to topics to speed up forum views