--- a/machina/apps/forum_tracking/handler.py
+++ b/machina/apps/forum_tracking/handler.py
@@ -11,6 +11,8 @@ from machina.core.loading import get_class
 Forum = get_model('forum', 'Forum')
 ForumReadTrack = get_model('forum_tracking', 'ForumReadTrack')
 TopicReadTrack = get_model('forum_tracking', 'TopicReadTrack')
+Topic = get_model('forum_conversation', 'Topic')
+Post = get_model('forum_conversation', 'Post')
 
 PermissionHandler = get_class('forum_permission.handler', 'PermissionHandler')
 
@@ -21,6 +23,7 @@ class TrackingHandler(object):
     in order to get only the forums which contain unread topics or the unread
     topics.
     """
+
     def __init__(self, request=None):
         self.request = request
         self.perm_handler = request.forum_permission_handler if request \
@@ -47,45 +50,57 @@ class TrackingHandler(object):
         Returns a list of unread topics for the given user from a given
         set of topics.
         """
-        unread_topics = []
-
-        # A user which is not authenticated will never see a topic as unread.
-        # If there are no topics to consider, we stop here.
-        if not user.is_authenticated() or topics is None or not len(topics):
-            return unread_topics
-
-        # A topic can be unread if a track for itself exists with a mark time that
-        # is less important than its update date.
         topic_ids = [topic.id for topic in topics]
-        topic_tracks = TopicReadTrack.objects.filter(topic__in=topic_ids, user=user)
-        tracked_topics = dict(topic_tracks.values_list('topic__pk', 'mark_time'))
 
-        if tracked_topics:
-            for topic in topics:
-                topic_last_modification_date = topic.last_post_on or topic.created
-                if topic.id in tracked_topics.keys() \
-                        and topic_last_modification_date > tracked_topics[topic.id]:
-                    unread_topics.append(topic)
+        # build query constraints
 
-        # A topic can be unread if a track for its associated forum exists with
-        # a mark time that is less important than its creation or update date.
-        forum_ids = [topic.forum_id for topic in topics]
-        forum_tracks = ForumReadTrack.objects.filter(forum_id__in=forum_ids, user=user)
-        tracked_forums = dict(forum_tracks.values_list('forum__pk', 'mark_time'))
+        in_topics = Q(id__in=topic_ids)
 
-        if tracked_forums:
-            for topic in topics:
-                topic_last_modification_date = topic.last_post_on or topic.created
-                if ((topic.forum_id in tracked_forums.keys() and topic.id not in tracked_topics) and
-                        topic_last_modification_date > tracked_forums[topic.forum_id]):
-                    unread_topics.append(topic)
+        updated_after_last_read_topic = (Q(tracks__user=user)
+                                         & (Q(tracks__mark_time__lt=F('last_post_on'))
+                                            | (Q(tracks__mark_time__lt=F('created')))))
 
-        # A topic can be unread if no tracks exists for it
-        for topic in topics:
-            if topic.forum_id not in tracked_forums and topic.id not in tracked_topics:
-                unread_topics.append(topic)
+        updated_after_last_read_forum = (Q(forum__tracks__user=user)
+                                         & (Q(forum__tracks__mark_time__lt=F('last_post_on'))
+                                            | (Q(forum__tracks__mark_time__lt=F('created')))))
 
-        return list(set(unread_topics))
+        updated_before_last_read_topic = (Q(tracks__user=user)
+                                          & (Q(tracks__mark_time__gte=F('last_post_on'))
+                                             | (Q(tracks__mark_time__gte=F('created')))))
+
+        untracked = (~Q(tracks__user=self.request.user) & ~Q(forum__tracks__user=user))
+
+        untracked_ids = Topic.approved_objects.filter(in_topics & untracked).values_list('id', flat=True)
+
+        not_tracked = Q(id__in=untracked_ids)
+
+        # run query
+        unread_topics = Topic.approved_objects.filter(in_topics & ((updated_after_last_read_topic
+                                                           | (updated_after_last_read_forum
+                                                              & ~updated_before_last_read_topic))
+                                                          | not_tracked))
+
+        return unread_topics
+
+    def get_oldest_unread_post(self, topic, user):
+
+        if not user.is_authenticated() or topic is None:
+            return None
+
+        read_posts = Post.approved_objects.filter(Q(topic=topic)
+                                                  & Q(topic__tracks__mark_time__gte=F('created'))
+                                                  ).values_list("id", flat=True)
+
+        unread_posts = Post.approved_objects.filter(Q(topic=topic)
+                                                    & (Q(topic__tracks__mark_time__lt=F('created'))
+                                                       | (Q(topic__forum__tracks__mark_time__lt=F('created'))
+                                                          & ~Q(id__in=read_posts)))
+                                                    ).order_by('created')[:1]
+
+        if unread_posts:
+            return unread_posts[0].pk
+        else:
+            return None
 
     def mark_forums_read(self, forums, user):
         """
@@ -119,7 +134,7 @@ class TrackingHandler(object):
             forum_track = None
 
         if forum_track is None \
-                or (topic.last_post_on and forum_track.mark_time < topic.last_post_on):
+            or (topic.last_post_on and forum_track.mark_time < topic.last_post_on):
             topic_track, created = TopicReadTrack.objects.get_or_create(topic=topic, user=user)
             if not created:
                 topic_track.save()  # mark_time filled
@@ -133,8 +148,8 @@ class TrackingHandler(object):
 
             forum_topic_tracks = TopicReadTrack.objects.filter(topic__forum=forum, user=user)
             if not unread_topics.exists() and (
-                    forum_track is not None or
-                    forum_topic_tracks.count() == forum.topics.filter(approved=True).count()):
+                        forum_track is not None or
+                        forum_topic_tracks.count() == forum.topics.filter(approved=True).count()):
                 # The topics that are marked as read inside the forum for the given user will be
                 # deleted while the forum track associated with the user must be created or updated.
                 # This is done only if there are as many topic tracks as approved topics in case