--- a/machina/apps/forum_tracking/views.py
+++ b/machina/apps/forum_tracking/views.py
@@ -5,6 +5,8 @@ from __future__ import unicode_literals
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.core.urlresolvers import reverse
+from django.db.models import F
+from django.db.models import Q
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404
 from django.utils.decorators import method_decorator
@@ -79,7 +81,7 @@ class MarkForumsReadView(TemplateView):
 
 
 class MarkTopicsReadView(
-        PermissionRequiredMixin, SingleObjectTemplateResponseMixin, BaseDetailView):
+    PermissionRequiredMixin, SingleObjectTemplateResponseMixin, BaseDetailView):
     """ Marks a set of topics as read. """
 
     model = Forum
@@ -129,9 +131,40 @@ class UnreadTopicsView(ListView):
     def get_queryset(self):
         forums = self.request.forum_permission_handler.get_readable_forums(
             Forum.objects.all(), self.request.user)
-        topics = Topic.objects.filter(forum__in=forums)
-        topics_pk = map(lambda t: t.pk, track_handler.get_unread_topics(topics, self.request.user))
-        return Topic.approved_objects.filter(pk__in=topics_pk).order_by('-last_post_on')
+
+        # build query constraints
+
+        in_forums = Q(forum__in=forums)
+
+        updated_after_last_read_topic = (Q(tracks__user=self.request.user)
+                                         & (Q(tracks__mark_time__lt=F('last_post_on'))
+                                            | (Q(tracks__mark_time__lt=F('created')))))
+
+        updated_after_last_read_forum = (Q(forum__tracks__user=self.request.user)
+                                         & (Q(forum__tracks__mark_time__lt=F('last_post_on'))
+                                            | (Q(forum__tracks__mark_time__lt=F('created')))))
+
+        updated_before_last_read_topic = (Q(tracks__user=self.request.user)
+                                          & (Q(tracks__mark_time__gte=F('last_post_on'))
+                                             | (Q(tracks__mark_time__gte=F('created')))))
+
+        untracked = ~(Q(tracks__user=self.request.user) | Q(forum__tracks__user=self.request.user))
+
+        untracked_ids = Topic.approved_objects.filter(in_forums & untracked).values_list('id', flat=True)
+
+        not_tracked = Q(id__in=untracked_ids)
+
+        # run query
+        topics = Topic.approved_objects.filter(in_forums & (updated_after_last_read_topic
+                                                            | (updated_after_last_read_forum
+                                                               & ~updated_before_last_read_topic)
+                                                            | not_tracked)) \
+            .order_by('-last_post_on').distinct().select_related('poster',
+                                                                   'forum',
+                                                                   'last_post',
+                                                                   'last_post__poster')
+
+        return topics
 
     @method_decorator(login_required)
     def dispatch(self, request, *args, **kwargs):