--- a/machina/templates/machina/forum_conversation/topic_list.html
+++ b/machina/templates/machina/forum_conversation/topic_list.html
@@ -25,7 +25,8 @@
                 <i class="fa {% if topic.is_sticky %}fa-lightbulb-o{% elif topic.is_announce %}fa-info{% else %}fa-dot-circle-o{% endif %} fa-2x"></i>
               </td>
               <td class="topic-name">
-                <a href="{% url 'forum_conversation:topic' topic.forum.slug topic.forum.pk topic.slug topic.pk %}" class="topic-name-link">{{ topic.subject }}</a>{% if topic.is_locked %}&nbsp;<i class="fa fa-times-circle locked-indicator" title="{% trans 'This topic is locked' %}"></i>{% endif %}
+                {% if topic in unread_topics or force_all_unread %}{% get_oldest_unread_post topic request.user as oldest_unread_post %}{% endif %}
+                <a href="{% url 'forum_conversation:topic' topic.forum.slug topic.forum.pk topic.slug topic.pk %}{% if oldest_unread_post != None %}?post={{ oldest_unread_post }}#{{ oldest_unread_post }}{% endif %}" class="topic-name-link">{{ topic.subject }}</a>{% if topic.is_locked %}&nbsp;<i class="fa fa-times-circle locked-indicator" title="{% trans 'This topic is locked' %}"></i>{% endif %}
                 <div>
                   <div class="topic-created">
                   {% if topic.poster %}