The code in `apps/forum/views.py` selects the latest topics ordered by type and last_post_on which is in forum_conversation_topic. Adding an index allows the database to use a btree index scan instead of heapsort. This yields a massive performance improvement in query time. Sample below in forum with 1.1m posts

```
Executed SQL
    SELECT ••• FROM "forum_conversation_topic" LEFT OUTER JOIN "accounts_user" ON ("forum_conversation_topic"."poster_id" = "accounts_user"."id") LEFT OUTER JOIN "forum_conversation_post" ON ("forum_conversation_topic"."last_post_id" = "forum_conversation_post"."id") LEFT OUTER JOIN "accounts_user" T5 ON ("forum_conversation_post"."poster_id" = T5."id") WHERE ("forum_conversation_topic"."forum_id" = 9 AND NOT ("forum_conversation_topic"."type" = 2) AND NOT ("forum_conversation_topic"."approved" = false)) ORDER BY "forum_conversation_topic"."type" DESC, "forum_conversation_topic"."last_post_on" DESC LIMIT 20
Time
    334.56873893737793 ms
Database
    default

```

Before indexing
```
QUERY PLAN
Limit  (cost=219444.88..219444.93 rows=20 width=1878) (actual time=369.051..369.065 rows=20 loops=1)
  ->  Sort  (cost=219444.88..219561.82 rows=46779 width=1878) (actual time=369.051..369.064 rows=20 loops=1)
        Sort Key: forum_conversation_topic.type DESC, forum_conversation_topic.last_post_on DESC
        Sort Method: top-N heapsort  Memory: 47kB
        ->  Hash Left Join  (cost=908.51..218200.11 rows=46779 width=1878) (actual time=6.636..329.387 rows=46749 loops=1)
              Hash Cond: (forum_conversation_post.poster_id = t5.id)
              ->  Nested Loop Left Join  (cost=454.47..217102.85 rows=46779 width=1779) (actual time=3.586..298.115 rows=46749 loops=1)
                    ->  Hash Left Join  (cost=454.04..5389.46 rows=46779 width=207) (actual time=3.571..63.752 rows=46749 loops=1)
                          Hash Cond: (forum_conversation_topic.poster_id = accounts_user.id)
                          ->  Seq Scan on forum_conversation_topic  (cost=0.00..4292.21 rows=46779 width=108) (actual time=0.010..33.853 rows=46749 loops=1)
                                Filter: (approved AND (type <> 2) AND (forum_id = 9))
                                Rows Removed by Filter: 82265
                          ->  Hash  (cost=311.24..311.24 rows=11424 width=99) (actual time=3.526..3.526 rows=11424 loops=1)
                                Buckets: 16384  Batches: 1  Memory Usage: 1649kB
                                ->  Seq Scan on accounts_user  (cost=0.00..311.24 rows=11424 width=99) (actual time=0.006..1.763 rows=11424 loops=1)
                    ->  Index Scan using forum_conversation_post_pkey on forum_conversation_post  (cost=0.43..4.52 rows=1 width=1572) (actual time=0.004..0.004 rows=1 loops=46749)
                          Index Cond: (forum_conversation_topic.last_post_id = id)
              ->  Hash  (cost=311.24..311.24 rows=11424 width=99) (actual time=3.006..3.006 rows=11424 loops=1)
                    Buckets: 16384  Batches: 1  Memory Usage: 1649kB
                    ->  Seq Scan on accounts_user t5  (cost=0.00..311.24 rows=11424 width=99) (actual time=0.004..1.190 rows=11424 loops=1)
Planning time: 1.192 ms
Execution time: 369.159 ms
```

After indexing

```
QUERY PLAN
Limit  (cost=1.42..109.13 rows=20 width=1878) (actual time=0.018..0.334 rows=20 loops=1)
  ->  Nested Loop Left Join  (cost=1.42..251930.88 rows=46779 width=1878) (actual time=0.018..0.332 rows=20 loops=1)
        ->  Nested Loop Left Join  (cost=1.13..237294.42 rows=46779 width=1779) (actual time=0.016..0.283 rows=20 loops=1)
              ->  Nested Loop Left Join  (cost=0.70..25581.03 rows=46779 width=207) (actual time=0.013..0.190 rows=20 loops=1)
                    ->  Index Scan Backward using forum_conversation_topic_type_last_post_on_idx on forum_conversation_topic  (cost=0.42..10034.59 rows=46779 width=108) (actual time=0.009..0.138 rows=20 loops=1)
                          Filter: (approved AND (type <> 2) AND (forum_id = 9))
                          Rows Removed by Filter: 37
                    ->  Index Scan using accounts_user_pkey on accounts_user  (cost=0.29..0.32 rows=1 width=99) (actual time=0.002..0.002 rows=1 loops=20)
                          Index Cond: (forum_conversation_topic.poster_id = id)
              ->  Index Scan using forum_conversation_post_pkey on forum_conversation_post  (cost=0.43..4.52 rows=1 width=1572) (actual time=0.004..0.004 rows=1 loops=20)
                    Index Cond: (forum_conversation_topic.last_post_id = id)
        ->  Index Scan using accounts_user_pkey on accounts_user t5  (cost=0.29..0.30 rows=1 width=99) (actual time=0.002..0.002 rows=1 loops=20)
              Index Cond: (forum_conversation_post.poster_id = id)
Planning time: 1.025 ms
Execution time: 0.403 ms
```