@ellmetha 2017-04-15
Thanks @norman-tan, this looks great! :wink:  This change should involve a Django migration. Could you generate this missing migration?